Move the sidebar into the details panel below the timeline header
Fixes #1061
