--- src/components/app/ProfileViewerContainer.css
@@ -1,21 +1,0 @@
-.ProfileViewerContainer .layout-pane > * {
-  box-sizing: border-box;
-  width: 100%;
-  height: 100%;
-}
-
-.ProfileViewerContainer .layout-pane:not(.layout-pane-primary) {
-  max-width: 250px;
-}
-
-/* overriding defaults from splitter-layout */
-.ProfileViewerContainer > .layout-splitter {
-  background: var(--grey-10); /* Same background as sidebars */
-  border-left: 1px solid var(--grey-40);
-}
-
-.ProfileViewerContainer > .layout-splitter:hover {
-  background: var(--grey-40); /* same as the border above */
-}
-
-
