+++ src/components/app/Root.js
@@ -12,7 +12,7 @@ import {
   retrieveProfileFromStore,
   retrieveProfileOrZipFromUrl,
 } from '../../actions/receive-profile';
-import ProfileViewerContainer from './ProfileViewerContainer';
+import ProfileViewer from './ProfileViewer';
 import ZipFileViewer from './ZipFileViewer';
 import Home from './Home';
 import { getView } from '../../reducers/app';
@@ -201,7 +201,7 @@ class ProfileViewWhenReadyImpl extends PureComponent<ProfileViewProps> {
         // The data is now loaded. This could be either a single profile, or a zip file
         // with multiple profiles. Only show the ZipFileViewer if the data loaded is a
         // Zip file, and there is no stored path into the zip file.
-        return hasZipFile ? <ZipFileViewer /> : <ProfileViewerContainer />;
+        return hasZipFile ? <ZipFileViewer /> : <ProfileViewer />;
       case 'ROUTE_NOT_FOUND':
       default:
         // Assert with Flow that we've handled all the cases, as the only thing left
