+++ src/test/components/Details.test.js
@@ -6,14 +6,14 @@
 import * as React from 'react';
 import { shallowWithStore } from '../fixtures/enzyme';
 
-import ProfileViewerContainer from '../../components/app/ProfileViewerContainer';
+import Details from '../../components/app/Details';
 import { changeSelectedTab } from '../../actions/app';
 import { storeWithProfile } from '../fixtures/stores';
 import { getProfileFromTextSamples } from '../fixtures/profiles/make-profile';
 
 import type { TabSlug } from '../../types/actions';
 
-describe('app/ProfileViewerContainer', function() {
+describe('app/DetailsContainer', function() {
   const { profile } = getProfileFromTextSamples(`
     A A A
     B B B
@@ -30,19 +30,19 @@ describe('app/ProfileViewerContainer', function() {
     'calltree', // 'calltree' not first on purpose
   ];
 
-  it('renders an initial view', () => {
+  it('renders an initial view with the right panel', () => {
     const store = storeWithProfile(profile);
     // dive() will shallow-render the wrapped component
-    const view = shallowWithStore(<ProfileViewerContainer />, store);
+    const view = shallowWithStore(<Details />, store);
     expect(view.dive()).toMatchSnapshot();
   });
 
   tabSlugs.forEach((tabSlug: TabSlug) => {
-    it(`renders initial views with tab ${tabSlug}`, () => {
+    it(`renders an initial view with the right panel for tab ${tabSlug}`, () => {
       const store = storeWithProfile(profile);
       store.dispatch(changeSelectedTab(tabSlug));
 
-      const view = shallowWithStore(<ProfileViewerContainer />, store);
+      const view = shallowWithStore(<Details />, store);
       expect(view.dive()).toMatchSnapshot();
     });
   });
